+++ src/Repositories/Interfaces/UserRepository.php
@@ -10,6 +10,8 @@ interface UserRepository
 {
     public function createAndReturnUser(User $user): User;
 
+    public function getById(string $id): User;
+
     public function getByEmail(string $email): ?User;
 
     public function updateGoogleId(User $user, string $googleId): void;
